requirements.txt
# Core Dependencies
selenium>=4.27.1
undetected-chromedriver>=3.5.5
selenium-stealth>=1.0.6
beautifulsoup4>=4.12.3
lxml>=5.3.0

# HTTP & API
requests>=2.31.0
aiohttp>=3.13.3

# Configuration & Environment
python-dotenv>=1.0.1
pyyaml>=6.0.2

# Database
mysql-connector-python>=9.2.0

# Scheduling
schedule>=1.2.2

# Browser Management
webdriver-manager>=4.0.2
humancursor>=1.1.0

# Utilities
certifi>=2024.12.14
charset-normalizer>=3.4.1
urllib3>=2.5.0
websocket-client>=1.8.0